Modularização do menu (matriz 4x4)
